% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmsviewer.R
\name{tdmsviewer}
\alias{tdmsviewer}
\title{Launch the tdmsviewer app}
\usage{
tdmsviewer(basedir = "~", sqlitePath = "~/sql.db", dev = F)
}
\arguments{
\item{basedir}{Root directory for TDMS files. Default '~'}

\item{sqlitePath}{Path to sqlite DB. Default '~/sql.db'}

\item{dev}{Boolean if using dev environment, loads from local directories}
}
\description{
Executing this function will launch the tdmsviewer application in
with a given folder of TDMS files
}
\examples{
\dontrun{
tdmsviewer(basedir='/data/dir')
}
}
