% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmsviewer.R
\name{tdmsviewer}
\alias{tdmsviewer}
\title{Launch the tdmsviewer app}
\usage{
tdmsviewer()
}
\description{
Executing this function will launch the tdmsviewer application in
the user's default web browser.
}
\examples{
\dontrun{
tdmsviewer()
}
}
\author{
Colin Diesh \email{dieshcol@msu.edu}
}
